Refactor DropEncryptRuleStatementUpdaterTest
